requirements.txt
Flask==3.0.0
biopython==1.85
numpy>=1.24.0
matplotlib>=3.6.0
reportlab>=4.0.0
networkx>=2.8
requests>=2.28.0
Pillow>=9.5.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
seaborn>=0.12.0
logomaker>=0.8